Cleanup: use actual buffer size in read()
phTmlUwb_spi_read() calls read() always with 4 buffer count
disregarding the actual buffer size(4200).

The current scheme works because both the kernel driver and
HAL assume the actual buffer is 4200 bytes, even if the given
length is 4.

Fix it to provide the actual buffer size to the kernel.
The kernel driver anyway doesn't check the count but it also
should be fixed.

Additionally, print errno on errors from read()/write()
instead of its return value.

Bug: 321604848
Test: uwb hal init

+++ halimpl/tml/phTmlUwb_spi.h
@@ -47,6 +47,6 @@
 void phTmlUwb_spi_close(void* pDevHandle);
 tHAL_UWB_STATUS phTmlUwb_spi_open_and_configure(pphTmlUwb_Config_t pConfig,
                                           void** pLinkHandle);
-int phTmlUwb_spi_read(void* pDevHandle, uint8_t* pBuffer, int nNbBytesToRead);
-int phTmlUwb_spi_write(void* pDevHandle, uint8_t* pBuffer, int nNbBytesToWrite);
+int phTmlUwb_spi_read(void* pDevHandle, uint8_t* pBuffer, size_t nNbBytesToRead);
+int phTmlUwb_spi_write(void* pDevHandle, uint8_t* pBuffer, size_t nNbBytesToWrite);
 int phTmlUwb_Spi_Ioctl(void* pDevHandle, phTmlUwb_ControlCode_t cmd, long arg);
